requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-groq
python-dotenv
streamlit
pypdf
